pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-base"
version = "0.1.0"
description = "Base framework for building MCP services with automatic handler discovery"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "MCP Base Contributors"}
]
keywords = ["mcp", "framework", "dependency-injection", "fastapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

# Core dependencies
dependencies = [
    "fastapi>=0.121.1",
    "uvicorn[standard]>=0.38.0",
    "pydantic>=2.12.4",
    "pyiv @ git+https://github.com/rl337/pyiv.git@main#egg=pyiv",  # Use main for now, will update to v0.2.0 when merged
    "prometheus-client>=0.20.0",  # Prometheus metrics
    "opentelemetry-api>=1.24.0",  # OpenTelemetry tracing
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "black>=25.11.0",
    "ruff>=0.14.4",
    "mypy>=1.18.2",
]

[tool.setuptools]
packages = ["mcp_base"]

[tool.black]
line-length = 100
target-version = ["py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["--strict-markers", "-v"]

